Add BTreeMap to the HASHMAP_ENTRY rule
Fixes #433
